optimize project structure
