+++ model/contact.py
@@ -1,7 +1,7 @@
 # @Author: Ackerven
 # @Mail: ackerven@cxmail.org
 # @Time: 14/11/2021 11:49
-# @File: contact.py
+# @File: model.py
 # OS: Ubuntu 20.04 LTS
 # SoftWare: PyCharm
 # @Copyright Copyright(C) 2021 Ackerven All rights reserved.
