+++ start.py
@@ -7,7 +7,7 @@
 # @Copyright Copyright(C) 2021 Ackerven All rights reserved.
 
 import yaml
-from user.console import console
+from view.console import console
 
 if __name__ == '__main__':
     config = yaml.load(open('config.yaml', 'r', encoding='utf-8'), Loader=yaml.FullLoader)
